requirements.txt
diffusers
peft
Pillow
opencv-python
numpy
safetensors
huggingface-hub
accelerate